Fix templates et template view
