+++ src/administrator/components/com_kunena/Controller/TemplateController.php
@@ -15,8 +15,10 @@ namespace Kunena\Forum\Administrator\Controller;
 defined('_JEXEC') or die();
 
 use Joomla\CMS\MVC\Controller\FormController;
-
-require_once __DIR__ . '/TemplatesController.php';
+use Joomla\CMS\Language\Text;
+use Joomla\CMS\Session\Session;
+use Joomla\Utilities\ArrayHelper;
+use Joomla\CMS\Filesystem\File;
 
 /**
  * Kunena Template Controller
@@ -25,4 +27,107 @@ require_once __DIR__ . '/TemplatesController.php';
  */
 class TemplateController extends FormController
 {
+	/**
+	 * Save template settings
+	 *
+	 * @return  void
+	 *
+	 * @since   Kunena 2.0
+	 *
+	 * @throws  Exception
+	 * @throws  null
+	 */
+	public function save()
+	{
+		$template = $this->app->input->get('templatename', '', 'cmd');
+		$menus    = $this->app->input->get('selections', [], 'array');
+		$menus    = ArrayHelper::toInteger($menus);
+
+		if (!Session::checkToken('post'))
+		{
+			$this->app->enqueueMessage(Text::_('COM_KUNENA_ERROR_TOKEN'), 'error');
+			$this->setRedirect(\KunenaRoute::_($this->baseurl, false));
+
+			return;
+		}
+
+		if (!$template)
+		{
+			$this->app->enqueueMessage(Text::_('COM_KUNENA_A_TEMPLATE_MANAGER_OPERATION_FAILED') . ': ' . Text::_('COM_KUNENA_A_TEMPLATE_MANAGER_TEMPLATE_NOT_SPECIFIED'));
+			$this->setRedirect(\KunenaRoute::_($this->baseurl, false));
+
+			return;
+		}
+
+		$this->_saveParamFile($template);
+
+		$this->app->enqueueMessage(Text::_('COM_KUNENA_A_TEMPLATE_MANAGER_CONFIGURATION_SAVED'));
+		$this->setRedirect(\KunenaRoute::_($this->baseurl, false));
+	}
+
+	/**
+	 * Apply
+	 *
+	 * @return  void
+	 *
+	 * @since   Kunena 2.0
+	 *
+	 * @throws  Exception
+	 * @throws  null
+	 */
+	public function apply()
+	{
+		$template = $this->app->input->get('templatename', '', 'cmd');
+		$menus    = $this->app->input->get('selections', [], 'array');
+		$menus    = ArrayHelper::toInteger($menus);
+
+		if (!Session::checkToken('post'))
+		{
+			$this->app->enqueueMessage(Text::_('COM_KUNENA_ERROR_TOKEN'), 'error');
+			$this->setRedirect(\KunenaRoute::_($this->baseurl, false));
+
+			return;
+		}
+
+		if (!$template)
+		{
+			$this->app->enqueueMessage(Text::_('COM_KUNENA_A_TEMPLATE_MANAGER_OPERATION_FAILED') . ': ' . Text::_('COM_KUNENA_A_TEMPLATE_MANAGER_TEMPLATE_NOT_SPECIFIED'));
+			$this->setRedirect(\KunenaRoute::_($this->baseurl, false));
+
+			return;
+		}
+
+		$this->_saveParamFile($template);
+
+		$this->app->enqueueMessage(Text::_('COM_KUNENA_A_TEMPLATE_MANAGER_CONFIGURATION_SAVED'));
+		$this->setRedirect(\KunenaRoute::_($this->baseurl . '&layout=edit&cid[]=' . $template, false));
+	}
+
+	/**
+	 * Method to restore the default settings of the template selected
+	 *
+	 * @return  void
+	 *
+	 * @since   Kunena 5.1
+	 *
+	 * @throws  Exception
+	 */
+	public function restore()
+	{
+		$template = $this->app->input->get('templatename', '', 'cmd');
+		$file     = KPATH_SITE . '/template/' . $template . '/config/params.ini';
+
+		if (file_exists($file))
+		{
+			$result = File::delete($file);
+
+			if ($result)
+			{
+				File::write($file, '');
+			}
+		}
+
+		$this->app->enqueueMessage(Text::_('COM_KUNENA_TEMPLATES_SETTINGS_RESTORED_SUCCESSFULLY'));
+		$this->setRedirect(\KunenaRoute::_($this->baseurl, false));
+	}
 }
